MAJOR CHANGE: Migrate to x86_64

+++ src/idt.h
@@ -2,5 +2,5 @@
 #include <stdint.h>
 
 void idt_init();
-void idt_set_gate(int n, uint32_t handler);
+void idt_set_gate(int n, uintptr_t handler);
